changed chart.zoom mechanism: moving z with perspective camera > scaling field with orthographic camera

+++ Assets/_scripts/scenes/track_play/Chart.cs
@@ -50,7 +50,7 @@ class Chart {
     [System.Serializable]
     public class Zoom {
         public float t;
-        public int d;
+        public float d;
         public float dur;
         public byte i;
     }
